Add tarball.WriteTar and tarball.WriteTargz
Gzipping before writing to the output makes it impossible for callers to
control the gzip implementation, but more importantly it makes it
impossible for callers to get access to the uncompressed tar stream,
which is necessary for computing the DiffID for OCI layers.

This preserves the behavior of WriteArchive, but deprecates it in favor
of WriteTar and WriteTargz, which also take a context.Context.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>