Start gunzipping ahead of time (#85)
Since we serialize installPackage, we just kind of spin doing nothing
once we've expanded all the APKs. For large APKs, we are currently
bottlenecked by gzip. We could cache the uncompressed tarballs, but that
would eat up a bunch of disk space (which is currently a pain point).

Instead, this starts decompressing as soon as we've finished expanding
APKs so that we can overlap that work with installing prerequisite
packages.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>