Install apk files lazily (#103)
This adds an alternative codepath where we can lazily install APK files
if the underlying filesystem impl supports it.

A filesystem implementation impl that supports that kind of thing is
being added to an internal apko package for now. We may want to move it
around (especially if I can come up with a nicer way to compose than
just forking memfs).

The piece that remains in go-apk will be an fs.FS implementation that
uses a ReadSeeker to seek around the APK's data section tar rather than
unpacking the data section to disk as individual files. Doing one Seek
is much faster than writing a bunch of temporary files to disk, reading
them back from disk, and then cleaning them up.

On a fairly large image build (go-1.20), this is about twice as fast as
the previous aproach.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>