Avoid leaking open files
I was trying to be clever by reusing open files, but it doesn't make a
huge difference in my testing, so we can just open/close them as needed.

We could probably do even better by having a single open file and
calling ReadAt under the hood, but let's stay simple for now.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>