Add global parsed index cache
This dedupes parsing the index (and the index itself) so we aren't
parsing the same index over and over again (and holding a bunch of
copies in memory).

Again, not the best, but this is the second of three things we're going
to dedupe in-process like this, mostly for the terraform provider, which
does a lot of simultaneous work that is very very similar and amenable
to this kind of deduplication.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>