Fix Replaces
This should have been a []string but was a string.

Also stop writing it to APKINDEX. This is only present in .PKGINFO, but
we also need to write it into the installed db.

Since it's not supposed to be in the APKINDEX, also start reading this
from the APK's .PKGINFO.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>