Generate actual APKs for install tests
This uses the public InstallPackages method, so it won't bypass a lot of
useful logic. Currently, a lot of the unit tests would fail if we
checked for duplicate files in the installed DB. I've commented that
check out for now in there so I can make the fix separately.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>