Avoid redundant work in solver
This implements two optimizations in the solver.

There are a couple places where we call sortPackages and take the 0th
item, then ignore the rest of the results. This is wasteful, since
sorting the entire slice takes log(n) times longer than finding the
minimum element in the slice. Where we used to do that, we now call
bestPackage() instead.

We were also spending a lot of time allocating maps that contained
origins that we have already seen. I've hoisted that code up to the
point where it makes most sense so that we can share that result across
multiple calls to bestPackage/sortPackages.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>