Rewrite solver to avoid installing overlapping providers
When we select a package, we want to disqualify anything that would
conflict with it before trying to resolve its dependencies.

For example, consider this dependency graph:

  foo-1.0.0 -> {bar, baz~1.0}

  bar-1.0.0 -> {baz-1.0.0}

  bar-2.0.0 -> {baz-2.0.0}

  baz-1.0.0 -> {}
  baz-1.0.1 -> {}
  baz-1.0.2 -> {}
  baz-2.0.0 -> {}

Let's say we want to install "foo".

Our heuristic for resolving the constraint with the least options would
see that we have only 2 options for "bar" and 3 options for "baz~1.0",
so we would resolve "bar" first.

Unfortunately, resolving "bar" would give us bar-2.0.0 because we prefer
the highest version. Selecting bar-2.0.0 would bring in baz-2.0.0, which
would conflict with our baz~1.0 constraint.

This is silly, because we can see the baz~1.0 constraint before we even
decide to resolve "bar".

After this change, we will disqualify anything that provides "baz" but
does not satisfy the "~1.0" version selector, which means bar-2.0.0
would be unselectable.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>

Fix some comments

wip

hmm

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>