Add a less expensive way to split APKs
We have a bunch of code that calls ExpandApk because it exists, but
that's terrible because ExpandApk is hyper-optimized for apko's use,
which means it's super expensive but caches very well.

Callers can instead use Split to avoid the expensive parts and just get
an APK that is split into its sections.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>