Invalidate in-process cache for local indexes
For a remote APKINDEX, we really don't want it to get updated between
builds, so we have this gross little global cache that gets used across
builds within the same process. This works pretty well.

However, if we're building local packages with a local index, we'd like
to have that get updated between builds so that we can depend on things
we built previously.

This change introduces modtime-based cache invalidation for a local
index so that it's possible to do this.

I'd like to clean this up and have a cache get passed in from the
caller, but today is not the day for that.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>