apk,signature: add support to verify RSA256 signed Index
This adds support to correctly parse signature stream that contains
multiple types of signatures. This also refactors signing to allow
trivially changing signature type, and to have multiple if so desired.

See code comments for lots of caveats.

Once all tools upgrade to this version of the module, signing can be
changed from RSA to RSA256.

Signed-off-by: Dimitri John Ledkov <dimitri.ledkov@chainguard.dev>